DynamoDB: transact_write_items() now validates empty keys on Update-queries
